Move string related functions from logging.h to string_util.h
Summary:
This is part of an effort to club all string related utility functions
into a common place, in string_util.
logging.h should contain only contain functions related to logging.
