+++ src.mk
@@ -134,7 +134,6 @@ LIB_SOURCES =                                                   \
   util/filter_policy.cc                                         \
   util/hash.cc                                                  \
   util/log_buffer.cc                                            \
-  util/logging.cc                                               \
   util/murmurhash.cc                                            \
   util/random.cc                                                \
   util/rate_limiter.cc                                          \
